完成inputUI

+++ PXE-X09/PXE-X09/Src/DataStruct/MacDefine.h
@@ -310,7 +310,7 @@ extern uint16 DELAY_SETTINGS_MAX;//960 20，384 8 ，240 5 ，259 5.396
 //------------------------   输入通道  ---------------------
 /*机型的Input通道数及机型的Input通道的EQ数量,主音量在MUSIC(DataStructInput)中,只使用一路*/
 extern uint8 Input_CH_MAX;
-extern uint8 Input_CH_Volume_MAX;
+extern uint16 Input_CH_Volume_MAX;
 
 //------------------------   输出通道  ---------------------
 /*机型的输出通道数及机型的输出通道的EQ数量*/
